requirements.txt
# Core dependencies
aiohttp>=3.8.0
websockets>=10.0
asyncio-mqtt>=0.11.0

# Telegram Bot
python-telegram-bot>=20.0

# Exchange SDKs
hyperliquid-python-sdk>=0.1.0
eth-account>=0.8.0

# Data handling
typing-extensions>=4.0.0
pandas>=1.5.0

# Configuration and validation
pydantic>=1.10.0
pydantic-settings>=2.0.0

# Encryption
cryptography>=3.4.8

# Logging
loguru>=0.7.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0